requirements.txt
async_timeout==5.0.1
BeautifulSoup==3.2.2
blobfile==3.0.0
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
checks==0.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.6
curio==1.6
cx_Oracle==8.3.0
Cython==3.1.3
defusedxml==0.7.1
docutils==0.22
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
grandalf==0.8
greenlet==3.2.4
h2==4.2.0
h5py==3.14.0
html5lib==1.1
HTMLParser==0.0.2
httpx_aiohttp==0.1.8
hypothesis==6.138.2
importlib_metadata==8.0.0
importlib_resources==6.5.2
ipython==9.3.0
ipywidgets==8.1.7
Jinja2==3.1.6
jnius==1.1.0
js==1.0
jsonschema==4.25.1
keyring==25.6.0
langchain_anthropic==0.3.19
langchain_astradb==0.6.0
langchain_aws==0.2.31
langchain_azure_ai==0.1.5
langchain_chroma==0.2.5
langchain_cohere==0.4.5
langchain_community==0.3.27
langchain_deepseek==0.1.4
langchain_elasticsearch==0.3.2
langchain_fireworks==0.3.0
langchain_google_genai==2.1.9
langchain_google_vertexai==2.0.28
langchain_groq==0.3.7
langchain_huggingface==0.3.1
langchain_ibm==0.3.16
langchain_milvus==0.2.1
langchain_mistralai==0.2.11
langchain_mongodb==0.7.0
langchain_neo4j==0.5.0
langchain_nvidia_ai_endpoints==0.3.16
langchain_ollama==0.3.6
langchain_openai==0.3.31
langchain_perplexity==0.1.2
langchain_pinecone==0.2.11
langchain_postgres==0.0.15
langchain_qdrant==0.2.0
langchain_together==0.3.1
langchain_weaviate==0.0.5
langchain_xai==0.2.5
langchainhub==0.1.21
langgraph_api==0.3.1
langsmith_pyo3==0.1.0rc5
lark==1.2.2
lxml_html_clean==0.4.2
matplotlib==3.10.5
mtrand==0.1
numexpr==2.11.0
openapi_pydantic==0.5.1
outcome==1.3.0.post0
Pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
pycryptodome==3.23.0
pygraphviz==1.14
PyInstaller==6.15.0
pyodide==0.0.2
pyOpenSSL==25.1.0
pyppeteer==2.0.0
pysqlcipher3==1.2.0
pytest==8.4.1
python-dotenv==1.1.1
pytz==2025.2
rapidfuzz==3.13.0
scipy==1.16.1
scipy_doctest==2.0.1
setuptools==80.9.0
setuptools_scm==9.2.0
simplejson==3.20.1
simsimd==6.5.1
slack_sdk==3.36.0
socksio==1.0.0
sounddevice==0.5.2
Sphinx==8.2.3
sqlcipher3==0.5.4
sse_starlette==3.0.2
starlette==0.47.2
tensorflow==2.20.0
threadpoolctl==3.6.0
tornado==6.5.1
urllib3_secure_extra==0.1.0
uvloop==0.21.0
websockets==15.0.1
xmlrpclib==1.0.1
xxhash==3.5.0
